# api/openapi.yaml
# ─────────────────────────────────────────────────────
# このファイルは API 仕様の Single Source of Truth。
# ここから Go の型/サーバ I/F と TypeScript の型を生成する。
# ─────────────────────────────────────────────────────
openapi: 3.0.3

info:
  title: go-todo API
  description: TodoアプリのHTTP API仕様
  version: 0.1.0

servers:
  - url: http://localhost:8080/api
    description: ローカル（Gin 直）

# タグ： Swagger UIのグルーピング用
tags:
  - name: auth
    description: 認証
  - name: todo
    description: Todo CRUD

# ─────────────────────────────────────────────────────
# paths: 各エンドポイントの定義
# ─────────────────────────────────────────────────────
paths:
  # ───── Todo CRUD ─────
  /todos:
    # ボディなし GET、認証必須
    get:
      tags: [todo]
      operationId: listTodos # 生成コードの関数名の元
      summary: ログインユーザーのTodo一覧
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
        '401':
          $ref: '#/components/responses/Unauthorized'

    # ボディあり POST、認証必須
    post:
      tags: [todo]
      operationId: createTodo
      summary: Todo を新規作成
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /todos/{id}:
    # 複数メソッドで共通のパスパラメータ
    parameters:
      - $ref: '#/components/parameters/TodoId'

    get:
      tags: [todo]
      operationId: getTodo
      summary: Todo を1件取得
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [todo]
      operationId: updateTodo
      summary: Todo を更新
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [todo]
      operationId: deleteTodo
      summary: Todo を削除
      security:
        - bearerAuth: []
      responses:
        '204':
          description: 削除成功 (レスポンスボディなし)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ───── Auth ─────
  /auth/signup:
    post:
      tags: [auth]
      operationId: signup
      summary: 新規ユーザー登録
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: 登録成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

  /auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: ログイン (JWT 発行)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: ログイン成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: ログアウト (現状はクライアント側のトークン破棄のみ。将来サーバ側でトークン無効化リストを実装する余地を残すため auth 必須としている)
      security:
        - bearerAuth: []
      responses:
        '204':
          description: ログアウト成功 (レスポンスボディなし)
        '401':
          $ref: '#/components/responses/Unauthorized'

  /auth/me:
    get:
      tags: [auth]
      operationId: getMe
      summary: 現在ログイン中のユーザー情報を取得
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

# ─────────────────────────────────────────────────────
# components: 再利用パーツ
# ─────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── 共通パラメータ ─────────────────────────────────
  parameters:
    TodoId:
      name: id
      in: path
      required: true
      description: Todo ID (UUID)
      schema:
        type: string
        format: uuid

  # ── 共通レスポンス (エラー類はここにまとめる) ──
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: "Conflict (例: メールアドレス重複)"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  # ── 型 (スキーマ) 定義 ─────────────────────────────
  schemas:
    # 共通のエラーレスポンス型
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Validation Error

    User:
      type: object
      required: [id, user_name, email]
      properties:
        id:
          type: string
          format: uuid
        user_name:
          type: string
        email:
          type: string
          format: email

    # エンティティ
    Todo:
      type: object
      required: [id, user_id, title, is_completed, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid # Go側でuuid.UUID, TS側でstringになる
        user_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        is_completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ── リクエスト DTO ────────────────────────────────
    CreateTodoRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true

    UpdateTodoRequest:
      # 全フィールド optional の部分更新 (PUT だが PATCH 的セマンティクス)
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
        is_completed:
          type: boolean

    SignupRequest:
      type: object
      required: [email, password, user_name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        user_name:
          type: string
          minLength: 1

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    # ── レスポンス DTO ────────────────────────────────
    LoginResponse:
      # Bearer 方式なので token を body で返す。frontend は localStorage に保存する想定
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: JWT トークン (Authorization ヘッダーに Bearer 形式で付与)
        user:
          $ref: '#/components/schemas/User'